+++ db/migrate/20190829145138_create_data_migrations.rb
@@ -1,0 +1,6 @@
+class CreateDataMigrations < ActiveRecord::Migration[5.2]
+  def change
+    return if table_exists? :data_migrations
+    create_table :data_migrations, primary_key: "version", id: :string
+  end
+end