+++ spec/mailers/claim_mailer_spec.rb
@@ -2,7 +2,7 @@ require "rails_helper"
 
 RSpec.describe ClaimMailer, type: :mailer do
   describe "#submitted" do
-    let(:claim) { create(:claim, :submittable) }
+    let(:claim) { create(:claim, :submittable, first_name: "Abraham", surname: "Lincoln") }
     let(:mail) { ClaimMailer.submitted(claim) }
 
     it "renders the headers" do
@@ -11,6 +11,7 @@ RSpec.describe ClaimMailer, type: :mailer do
     end
 
     it "renders the body" do
+      expect(mail.body.encoded).to match("Dear Abraham Lincoln,")
       expect(mail.body.encoded).to match("We've received your application to claim back your student loan repayments for the time you spent at #{claim.eligibility.claim_school.name}.")
       expect(mail.body.encoded).to match("Your unique reference is #{claim.reference}. You will need this if you contact us about your claim.")
     end